requirements.txt
asgiref==3.7.2
Django==4.2
django-4-jet==1.0.9
django-cors-headers==4.0.0
django-environ==0.10.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
